added more interfaces and refactored code to implement interface
