+++ src/main/java/com/group6/wishshare/domain/service/WishService.java
@@ -1,5 +1,6 @@
 package com.group6.wishshare.domain.service;
 
+import com.group6.wishshare.data.repository.WishRepository;
 import com.group6.wishshare.data.repository.WishRepositoryImpl;
 import com.group6.wishshare.domain.model.Wish;
 
@@ -7,31 +8,31 @@ import java.util.List;
 
 public class WishService {
 
-  WishRepositoryImpl wishRepositoryImpl;
+  WishRepository wishRepository;
 
-  public WishService(WishRepositoryImpl wishRepositoryImpl) {
-    this.wishRepositoryImpl = wishRepositoryImpl;
+  public WishService(WishRepository wishRepository) {
+    this.wishRepository = wishRepository;
   }
 
   public void createWish(String name, String link, String price, int id) {
     Wish wish = new Wish.WishBuilder().name(name).link(link).price(price).wishlistId(id).build();
 
-    wishRepositoryImpl.addWishToWishList(wish);
+    wishRepository.addWishToWishList(wish);
   }
 
   public List<Wish> getWishes(int id) {
-    return wishRepositoryImpl.getWishes(id);
+    return wishRepository.getWishes(id);
   }
 
   public Wish getWish(int id) {
-    return wishRepositoryImpl.getWish(id);
+    return wishRepository.getWish(id);
   }
 
   public void reserveWish(boolean isReserved, int id) {
-    wishRepositoryImpl.isReservedWish(isReserved, id);
+    wishRepository.isReservedWish(isReserved, id);
   }
 
   public void editWish(String name, String link, String price, int id) {
-    wishRepositoryImpl.editWish(name, link, price, id);
+    wishRepository.editWish(name, link, price, id);
   }
 }
